Crate serde_aux

serde-aux

A serde auxiliary library.

Modules

container_attributes

Contains helpers for the containers.

field_attributes

Contains helpers for the fields.

prelude

Prelude module, contains the most needed helpers from this library.

Macros

enum_number_declare

This contains both serialization and ser/deserialization of a enum into and from numbers. The reference implementation does not work if your enum has negative values. This enum_number handles this also.